llvm: Move composition execution to a separate file
